Edit : Major refactor to modularize classification code and improve streaming of GinsengDataset. A suite of new tests have been added for restructured code. The code should be organized for future addition of regressor based models or multiple instance learning models.

+++ justfile
@@ -10,7 +10,7 @@
   uv run pytest -v
 
 @docs:
-  rm -r site
+  if [ -d site ]; then rm -r site; fi
   uv run python docs/parse.py
   uv run zensical build
